+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 
 # Changelog
 
+This package strictly follows [Semantic Versioning](https://semver.org).
+
 ## v1.9.0 (2022-10-23)
 
 ### Features
@@ -102,4 +104,4 @@
 
 ## v1.0.0 (2022-06-20)
 
-First major release. This package strictly follows [Semantic Versioning](https://semver.org).
+First major release.